Added grid based placing system
The player can now place cubes and the area where the cube is placed is highlighted (this still needs some work to make sure these are in sync). Additionally there were changes made to the character's camera movement - this was locked to no longer allow 360 flips in the vertical direction. Furthermore the cursor is now locked to the game which allows for placing of blocks. Grid size can be modified through  the unity attributes panel.